feat(devices): device service, V2 db context wiring, and initial schema migration
Adds the device registry service (DeviceService, IDeviceHandler, DeviceHandlerBase,
DeviceContext, DeviceCmdResult, all attributes) and root handler scaffolding.
Wires OosDbContext into the bootstrapper with unversioned keys (permanent) and
moves ApplicationDbContext to LEGACY: prefixed keys. Adds OosDbContextFactory for
EF design-time tooling. Fixes OrderStage self-referencing relationship config,
promotes OrderNumber to long with a unique index, caps DeviceType at 100 chars,
and generates the initial V2 schema migration.

+++ Data/DataModels/V2/Ordering/OrderHeader.cs
@@ -21,7 +21,7 @@ public class OrderHeader
     /// The human-readable order number, reset daily. Use <see cref="OrderDate"/> for
     /// full uniqueness when referencing historical orders.
     /// </summary>
-    public int OrderNumber { get; set; }
+    public long OrderNumber { get; set; }
 
     /// <summary>
     /// The date and time the order was placed, including local timezone offset.
